organização do context para termos
