+++ src/contexts/TermsContext/types.ts
@@ -1,0 +1,10 @@
+export interface TermsContent {
+    id: number;
+    name: string;
+    slug: string;
+    count: number;
+}
+
+export interface TermsContextType {
+    categories: TermsContent[];
+}
