+++ src/hooks/useFormattedDate.tsx
@@ -2,13 +2,12 @@ import { useMemo } from "react";
 
 const useFormattedDate = (dateString?: string): string => {
   return useMemo(() => {
-    if (!dateString) return ""; // Retorna uma string vazia se dateString for undefined
+    if (!dateString) return "";
 
     const date = new Date(dateString);
 
-    // Verifica se a data é válida
     if (isNaN(date.getTime())) {
-      return ""; // Retorna uma string vazia se a data não for válida
+      return "";
     }
 
     const options: Intl.DateTimeFormatOptions = {
@@ -20,7 +19,6 @@ const useFormattedDate = (dateString?: string): string => {
       hour12: false,
     };
 
-    // Formata a data para o padrão brasileiro e remove " às" se existir
     return date.toLocaleDateString("pt-BR", options).replace(" às", " às");
   }, [dateString]);
 };
